+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/BaseRichTextContentPartRender.cs
@@ -3,13 +3,13 @@ using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
 using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
-public abstract class RichTextContentRender : IRichTextContentPartRenderer
+public abstract class BaseRichTextContentPartRender : IRichTextContentPartRenderer
 {
     private readonly RichTextNodeType _nodeType;
 
-    protected RichTextContentRender(RichTextNodeType nodeType)
+    protected BaseRichTextContentPartRender(RichTextNodeType nodeType)
     {
         _nodeType = nodeType;
     }
