+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/TextRenderer.cs
@@ -3,10 +3,11 @@ using System.Text;
 using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
 using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Options;
 
 namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
-public class TextRenderer : RichTextContentRender
+public class TextRenderer : BaseRichTextContentPartRender
 {
     private readonly TextRendererOptions _textRendererOptions;
