+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/RichTextRenderer.cs
@@ -4,6 +4,9 @@ using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
 namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models;
 
+/// <summary>
+/// Parent class to render <see chref="IRichTextContent"/> RichTextContent
+/// </summary>
 public class RichTextRenderer : IRichTextRenderer, IRichTextContentPartRendererCollection
 {
     private readonly List<IRichTextContentPartRenderer> _renderers;
@@ -13,9 +16,19 @@ public class RichTextRenderer : IRichTextRenderer, IRichTextContentPartRendererC
         _renderers = renderers.ToList();
     }
 
+    /// <summary>
+    /// Finds matching renderer for the given content, based on the content's node type
+    /// </summary>
+    /// <param name="content">Content to find renderer for</param>
+    /// <returns>Matching part renderer for content (or null if not found)</returns>
     public IRichTextContentPartRenderer? GetRendererForContent(IRichTextContent content)
     => _renderers.FirstOrDefault(renderer => renderer.Accepts(content));
 
+    /// <summary>
+    /// Converts content to HTML string
+    /// </summary>
+    /// <param name="content">Content to convert</param>
+    /// <returns>Content converted to HTML string (including tags, classes, etc.)</returns>
     public string ToHtml(IRichTextContent content)
     {
         var stringBuilder = new StringBuilder();
