+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Options/MarkOptions.cs
@@ -1,0 +1,22 @@
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Options;
+
+public class MarkOptions
+{
+    /// <summary>
+    /// What mark this options is for (e.g. "bold" or "underline")
+    /// </summary>
+    /// <value></value>
+    public required string Mark { get; init; }
+
+    /// <summary>
+    /// What HTML tag should be generated (e.g. "strong")
+    /// </summary>
+    /// <value></value>
+    public required string HtmlTag { get; init; }
+
+    /// <summary>
+    /// Classes to add to generated tag
+    /// </summary>
+    /// <value></value>
+    public string? Classes { get; init; }
+}
