+++ src/Dfe.PlanTech.Infrastructure.Contentful/Helpers/ContentfulSetup.cs
@@ -4,6 +4,7 @@ using Contentful.Core.Configuration;
 using Dfe.PlanTech.Application.Persistence.Interfaces;
 using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 using Dfe.PlanTech.Infrastructure.Contentful.Persistence;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -37,7 +38,7 @@ public static class ContentfulSetup
 
     public static IServiceCollection SetupRichTextRenderer(this IServiceCollection services)
     {
-        var contentRendererType = typeof(RichTextContentRender);
+        var contentRendererType = typeof(BaseRichTextContentPartRender);
         var richTextPartRenderers = contentRendererType.Assembly.GetTypes()
                                                                 .Where(IsContentRenderer(contentRendererType));
